refactor: [#1282] move BanService to udp-tracker-core package
