+++ src/bootstrap/app.rs
@@ -35,8 +35,8 @@ use tracing::instrument;
 use super::config::initialize_configuration;
 use crate::bootstrap;
 use crate::container::AppContainer;
+use crate::packages::udp_tracker_core::services::banning::BanService;
 use crate::packages::{http_tracker_core, udp_tracker_core};
-use crate::servers::udp::server::banning::BanService;
 use crate::servers::udp::server::launcher::MAX_CONNECTION_ID_ERRORS_PER_IP;
 use crate::shared::crypto::ephemeral_instance_keys;
 use crate::shared::crypto::keys::{self, Keeper as _};
