+++ src/packages/tracker_api_core/statistics/services.rs
@@ -5,8 +5,9 @@ use packages::tracker_api_core::statistics::metrics::Metrics;
 use tokio::sync::RwLock;
 use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
 
-use crate::packages::{self, http_tracker_core, udp_tracker_core};
-use crate::servers::udp::server::banning::BanService;
+use crate::packages::udp_tracker_core::services::banning::BanService;
+use crate::packages::udp_tracker_core::{self};
+use crate::packages::{self, http_tracker_core};
 
 /// All the metrics collected by the tracker.
 #[derive(Debug, PartialEq)]
@@ -83,8 +84,8 @@ mod tests {
 
     use crate::packages::tracker_api_core::statistics::metrics::Metrics;
     use crate::packages::tracker_api_core::statistics::services::{get_metrics, TrackerMetrics};
+    use crate::packages::udp_tracker_core::services::banning::BanService;
     use crate::packages::{http_tracker_core, udp_tracker_core};
-    use crate::servers::udp::server::banning::BanService;
     use crate::servers::udp::server::launcher::MAX_CONNECTION_ID_ERRORS_PER_IP;
 
     pub fn tracker_configuration() -> Configuration {
