+++ src/servers/udp/server/mod.rs
@@ -6,7 +6,6 @@ use thiserror::Error;
 
 use super::RawRequest;
 
-pub mod banning;
 pub mod bound_socket;
 pub mod launcher;
 pub mod processor;
